CarIdentifier moving connectivity observer into network module
