+++ app/src/main/java/com/tomerpacific/caridentifier/composable/CarInformation.kt
@@ -37,7 +37,7 @@ import com.tomerpacific.caridentifier.R
 import com.tomerpacific.caridentifier.concatenateCarMakeAndModel
 import com.tomerpacific.caridentifier.model.CarDetails
 import com.tomerpacific.caridentifier.model.MainViewModel
-import com.tomerpacific.caridentifier.network.NO_INTERNET_CONNECTION_ERROR
+import com.tomerpacific.caridentifier.data.network.NO_INTERNET_CONNECTION_ERROR
 
 @Composable
 fun Details(mainViewModel: MainViewModel, serverError: State<String?>) {
